Switch to Integer ID for the history entry
git-svn-id: svn://localhost/jahia/trunk@9465 69c369ed-ed09-0410-a4a1-e19a46c23626

+++ war/src/main/webapp/WEB-INF/var/db/sql/schema/derby/jahia-schema.sql
@@ -20,7 +20,7 @@
     drop table jahia_version;
 
     create table jahia_contenthistory (
-        id bigint not null,
+        id integer not null,
         entry_action varchar(255),
         entry_date timestamp,
         message varchar(255),
